+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 - New problems about rules of squares
 - Mark entries younger than 2 weeks as `[NEW!]`
 - Trying to add a site description for SEO
+- New `symbols::get_drawable()` for non-alphabetic symbols
 
 ### Changed
 - Math API for creating `Polynomials` is simplified: `&t1 + &t2` instead of `t1.and(&t2)`